Ep 23

+++ app/Http/Controllers/JobController.php
@@ -3,6 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Job;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class JobController extends Controller
 {
@@ -69,6 +72,8 @@ class JobController extends Controller
     }
 
     public function update(Job $job){
+        Gate::authorize('edit-job', $job);
+        
         $title = request('title');
         $salary = trim(request('salary'));
         
@@ -94,7 +99,8 @@ class JobController extends Controller
         return redirect('/jobs/' . $job->id);
     }
 
-    public function destroy(Job $job){   
+    public function destroy(Job $job){  
+        Gate::authorize('edit-job', $job); 
         $job->delete();
         return redirect('/jobs');
     }
